+++ cuqi/experimental/__init__.py
@@ -1,0 +1,2 @@
+""" Experimental module for testing new features and ideas. """
+from . import mcmc
