+++ cuqi/experimental/mcmc/__init__.py
@@ -1,3 +1,5 @@
+""" Re-implementation of sampler module in a more object oriented way. """
+
 from ._sampler import SamplerNew, ProposalBasedSamplerNew
 from ._langevin_algorithm import MALANew
 from ._mh import MHNew
