+++ cuqi/experimental/mcmc/_mh.py
@@ -1,6 +1,6 @@
 import numpy as np
 import cuqi
-from cuqi.mcmc import ProposalBasedSamplerNew
+from cuqi.experimental.mcmc import ProposalBasedSamplerNew
 from cuqi.array import CUQIarray
